requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
google-generativeai==0.7.2
cachetools==5.3.3
feedparser==6.0.11
